refactor(codegen): extrai analysis/types.rs de func.rs
Folha sem deps cruzadas. Move sanitize_symbol, infer_expr_ty,
infer_abi_member_ty, ts_type_to_val_ty para
crates/rts-codegen/src/codegen/lower/analysis/types.rs.

func.rs: 9603 -> 9385 linhas. Comportamento preservado, suite verde.
Primeira fatia do refator do branch (issue interna sobre tamanho do
func.rs).
