+++ crates/rts-codegen/src/codegen/lower/analysis/mod.rs
@@ -1,0 +1,5 @@
+//! Analise estatica auxiliar do codegen.
+//!
+//! Submodulos folha (sem deps cruzadas entre si) extraidos de `func.rs`.
+
+pub mod types;
